Fix color icon loading, relocate gi_composites shim
