+++ lutris/gui/widgets/gi_composites.py
